requirements.txt
websockets==12.0
boto3==1.35.0
python-dotenv==1.0.1
ujson==5.10.0
